+++ src/selectors/archanova.js
@@ -36,14 +36,44 @@ import { firebaseRemoteConfig } from 'services/firebase';
 
 // selectors
 import { useActiveAccount, useRootSelector } from 'selectors';
+import { archanovaAccountIdSelector } from 'selectors/accounts';
+import { assetsBalancesPerAccountSelector } from 'selectors/balances';
+import { collectiblesPerAccountSelector } from 'selectors/collectibles';
 
 // types
 import type { RootReducerState, Selector } from 'reducers/rootReducer';
 import type { ArchanovaWalletStatus } from 'models/ArchanovaWalletStatus';
+import type { AssetBalancesPerAccount, WalletAssetsBalances } from 'models/Balances';
+import type { Collectible, CollectiblesStore } from 'models/Collectible';
 
 // local
 import { accountsSelector } from './selectors';
 import { archanovaAccountEthereumHistorySelector } from './history';
+
+
+/**
+ * Returns all wallet asset balances for archanova account.
+ */
+export const archanovaWalletAssetsBalancesSelector: Selector<WalletAssetsBalances> = createSelector(
+  archanovaAccountIdSelector,
+  assetsBalancesPerAccountSelector,
+  (archanovaAccountId: ?string, balancesPerAccount: AssetBalancesPerAccount): WalletAssetsBalances => {
+    if (!archanovaAccountId) return {};
+    return balancesPerAccount[archanovaAccountId]?.ethereum?.wallet ?? {};
+  },
+);
+
+/**
+ * Returns all collectibles for archanova account.
+ */
+export const archanovaCollectiblesSelector: Selector<Collectible[]> = createSelector(
+  archanovaAccountIdSelector,
+  collectiblesPerAccountSelector,
+  (archanovaAccountId: ?string, collectiblesPerAccount: CollectiblesStore): Collectible[] => {
+    if (!archanovaAccountId) return [];
+    return collectiblesPerAccount[archanovaAccountId]?.ethereum ?? [];
+  },
+);
 
 export const isArchanovaAccountDeployedSelector = ({ accounts: { data: accounts }, smartWallet }: RootReducerState) => {
   const archanovaWalletStatus: ArchanovaWalletStatus = getArchanovaWalletStatus(accounts, smartWallet);
